added fix to FixCommentPostRelationship

+++ db/migrate/20160519185834_fix_comment_post_relationship.rb
@@ -1,6 +1,10 @@
 class FixCommentPostRelationship < ActiveRecord::Migration
   def change
-    rename_column :comments, :commentable_id, :post_id
-    remove_column :comments, :commentable_type
+    if column_exists? :comments, :commentable_id
+      rename_column :comments, :commentable_id, :post_id
+    end
+    if column_exists? :comments, :commentable_type
+      remove_column :comments, :commentable_type
+    end
   end
 end
